requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
django-filter==25.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
django-environ==0.12.0
gunicorn==22.0.0
django-cors-headers==4.6.0
